requirements.txt
requests
selenium
huggingface_hub
python-dotenv